fixed ui, added my ratings, sorting by new/old

+++ app/Http/Controllers/RatingsController.php
@@ -36,6 +36,9 @@ class RatingsController extends Controller
     public function index()
     {
         //
+        $id = auth()->user()->id;
+        $ratings = ratings::all()->where('student_id', $id);
+        return view('ratings.user')->with('ratings', $ratings);
     }
 
     /**
@@ -125,8 +128,31 @@ class RatingsController extends Controller
      */
     public function edit($id)
     {
-        //
-    }
+        $rating = ratings::find($id);
+        $prof_id = $rating->professors['id'];
+
+        //    if(Auth::check()){
+        $ratings = ratings::all()->where('prof_id', $prof_id)->where('student_id', Auth::id());
+        if($rating->student_id == Auth::id()){
+            $student_id = auth()->user()->id;
+            $prof_id = $prof_id;
+            $professor = professors::find($prof_id);
+            $courses = Course::all()->where('university_id', $professor->university['id']);
+
+            $data=[
+                'courses' => $courses,
+                'rating' => $rating,
+                'student_id' => $student_id,
+                'prof_id' => $prof_id,
+                'professor' => $professor,
+            ];
+            return view('ratings.edit')->with($data);
+        } else{
+            return redirect('/professors/'.$prof_id)->with('success', 'One cannot edit what one did not create!')->with('rating', $rating);
+        }
+
+    }  
+    
 
     /**
      * Update the specified resource in storage.
@@ -138,6 +164,36 @@ class RatingsController extends Controller
     public function update(Request $request, $id)
     {
         //
+        $rating = ratings::find($id);
+        if($rating->student_id == Auth::id()){
+            try { 
+                $this->validate($request, [
+                'prof_id' => 'required',
+                'student_id' => 'required',
+                'course_id' => 'required',
+                'course_rating' => 'required',
+                'pteaching_rating' => 'required',
+                'pcharacter_rating' => 'required',
+                'pmastery_rating' => 'required',
+                'comment' => 'required',
+            ]);
+            // Update Rating
+            $rating = ratings::find($id);
+            $rating->student_id = $request->input('student_id');
+            $rating->prof_id = $request->input('prof_id');
+            $rating->course_id = $request->input('course_id');
+            $rating->course_rating = $request->input('course_rating');
+            $rating->pteaching_rating = $request->input('pteaching_rating');
+            $rating->pcharacter_rating = $request->input('pcharacter_rating');
+            $rating->pmastery_rating = $request->input('pmastery_rating');
+            $rating->comment = $request->input('comment');
+            $rating->save();
+            return redirect("/professors/{$request->input('prof_id')}")->with('success', 'Rating updated');
+            } catch(\Illuminate\Database\QueryException $ex){ 
+
+                return redirect("/professors/{$request->input('prof_id')}")->with('success', 'Most likely you have already rated this professor. If you didn\'t, please contact us at r8myprof@gmail.com to help u solve this issue.');
+            }
+        }
     }
 
     /**
@@ -150,7 +206,10 @@ class RatingsController extends Controller
     {
         //
         $rating = ratings::find($id);
-        $rating->delete();
-        return redirect('/professors/' . $rating->professors['id'])->with('success', 'Rating deleted');
+        if($rating->student_id == Auth::id()){
+            $rating = ratings::find($id);
+            $rating->delete();
+            return redirect('/professors/' . $rating->professors['id'])->with('success', 'Rating deleted');
+        }
     }
 }
